+++ backend/package.json
@@ -1,0 +1,7 @@
+{
+  "dependencies": {
+    "express": "^5.2.1",
+    "multer": "^2.3.0",
+    "sqlite3": "^6.0.1"
+  }
+}
